refactor(deploy-check): move tests module to a sibling file
The inline `#[cfg(test)] mod tests { ... }` block in main.rs is
now a sibling `tests.rs` reached via `mod tests;`. Same content,
same coverage; main.rs is no longer two-thirds test fixtures.
